+++ bindings/python/htcondor1/_deprecation.py
