+++ bindings/python/htcondor1/_lock.py
